+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Fix: #3 adjust `PasswordHasher` parameters for Raspberry Pi Zero W
 - Fix: catch `TimeoutError` in `Nest`
 - Fix: #3 authenticate user for `/system/db.sqlite3` and `/view/{key}`
+- Chore: change `/system/db.sqlite3` to `/db.sqlite3`
 
 ### `0.6`